Improve tests.
Split out test files.
Add missing histogram exposition test.
Add unittests for metric families.
